feat(ng): B2 — widen ng's coordinates to u64; delete the silent clamp
Spec S4: ng speaks one width, so nothing narrows, nothing is checked, and no
off-by-width bug is possible. RefSeq::fetch_into/fetch_raw_into and their three
impls, RefSeqError's coordinates, RepeatInterval/RegionSpan/SegmentOptions, the
coverage spans, and Bp are all u64. Ids stay u32 (they index a table, they are
not positions), and so do counts (min_copies) and scores.

The point of the exercise, deleted:

    contig_length: u32::try_from(contig_len).unwrap_or(u32::MAX)

On a contig above 4 Gb that error SILENTLY LIED about the very number it
existed to report, in built ng code, and nothing could catch it because
nothing could represent the right answer. Now it reports the real length, and
`a_request_beyond_u32_is_reported_verbatim_not_clamped` pins it.

RepeatInterval stays 0-based half-open (owner-agreed), against spec S4's
"1-based everywhere". It is a detector/slice type: its coordinates index the
byte slice find_tandem_repeats was handed, and seq[iv.start..iv.end] is what
every producer and consumer does. Making it 1-based would put a `- 1` at every
one of those sites to buy consistency with a genetic coordinate it never
expresses. The 1-based boundary belongs where the genetics starts -- Locus,
GenomeRegion -- which is where S4's payoff actually lands. Recorded on the type.

ng meets production's u32 at three seams (src/fasta/ is frozen), and every one
FAILS rather than folds -- a silent narrowing here would reintroduce exactly
the bug deleted above:
- ref_seq's WindowedRefSeq: new `narrow_for_fetcher` returns OutOfBounds, not
  a clamp, for a request the production fetcher cannot express.
- scan_windowed: narrows at the fetch with an `expect`, provably lossless
  because every coordinate derives from the fetcher's own u32 length; a
  truncated fetch would scan the WRONG WINDOW.
- The test differentials (ref_seq vs production's fetchers, scanner_parity and
  admission's build_loci bridge): narrow with `expect`, because a truncated
  coordinate would compare the wrong tract and pass while asserting nothing.

B2 also retired `Candidate`. A1 introduced it to hold "the four fields
build_loci reads, widened to u64", and its doc predicted the conversion would
become an identity but claimed the type would still earn its keep. It did not:
at u64 it was field-for-field identical to RepeatInterval, reachable only
through two no-op From impls, one of which carried a try_into().expect() and a
fallible_impl_from exemption for a narrowing that can no longer narrow. The
policy now works on RepeatInterval directly, and Admitted::bundled hands the
caller's own intervals back with no conversion -- which is what "handed back
verbatim" always claimed.

Every existing test is green unchanged, including the golden-catalog parity
(16/16) and the admission differential against production's real build_loci --
which is the evidence the widening changed no answer.

1827 lib tests; fmt/clippy/doc clean in scope. Only src/ng/ touched.

+++ src/ng/types.rs
@@ -53,14 +53,18 @@ impl BaseQual {
 
 /// A length in base pairs — the generic length currency both the SNP/indel and
 /// STR paths speak (only *repeat-unit* quantities carry the `Ssr` prefix). Here
-/// it measures a read's decoded length. Unconstrained — any `u32` is a legal
+/// it measures a read's decoded length. Unconstrained — any `u64` is a legal
 /// value, so the field is public and there is no checked constructor.
+///
+/// `u64` since B2 (spec §4): ng speaks one width, so nothing narrows, nothing is
+/// checked, and no off-by-width bug is possible. Ids stay `u32` — they index a
+/// table, they are not positions.
 #[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
-pub struct Bp(pub u32);
+pub struct Bp(pub u64);
 
 impl Bp {
     #[inline]
-    pub fn get(self) -> u32 {
+    pub fn get(self) -> u64 {
         self.0
     }
 }
